Clean up the code
